requirements.txt
dill
flatten_json
gradio
grequests
gseapy
hdbscan
hyperopt
imbalanced_learn
matplotlib
numpy
pandas
protobuf
pydeseq2
pymongo
pytest
Requests
scikit_learn
scipy
seaborn
shap
tqdm
umap_learn
sphinx
sphinx-automodapi
sphinx-rtd-theme